订单支付

+++ app/Http/Controllers/UserAddressesController.php
@@ -36,11 +36,15 @@ class UserAddressesController extends Controller
 
     public function edit(UserAddress $user_address)
     {
+        $this->authorize('own', $user_address);
+
         return view('user_addresses.create_and_edit',['address' => $user_address]);
     }
 
     public function update(UserAddress $user_address,Request $request)
     {
+        $this->authorize('own', $user_address);
+
         $user_address->update($request->only([
                                                  'province',
                                                  'city',
@@ -56,6 +60,8 @@ class UserAddressesController extends Controller
 
     public function destroy(UserAddress $user_address)
     {
+        $this->authorize('own', $user_address);
+
         $user_address->delete();
 
         return [];
